requirements.txt
Flask==3.0.2
Flask-MySQL==1.5.1
Flask-SQLAlchemy==3.1.0
PyJWT==2.8.0
gunicorn==21.2.0
bcrypt==4.0.1
python-dotenv==1.0.0